@peterbell10 2025-09-04
@stadlmax I'm still concerned about all the manual reference counting, so I took the liberty of pushing to your branch to use RAII. I had hoped we could use the pybind11 API itself, but after trying it out it seems like the exception handling is adversely effecting the latency. So this is a hybrid solution using `py::object` to handle reference counting but still calling the CPython API directly. 

On my machine this gives identical latency figures but IMO is less error prone. Let me know what you think.