@stadlmax 2025-09-05
I definitely share some of the concerns but also ran into the issue of exposing too much latency elsewhere if one indeed relies more on pybind11 itself. Relying on RAII to simplify the clean-up, however, indeed seems like an elegant solution which likely is useful in other native extensions, too. Thanks for being persistent and picking this up on yourself 👍🏽