@ThomasRaoux 2025-09-09
I'm getting some regression due to this PR with this error:
```
TypeError: failed to specialize argument of type: numpy.float64
```

@peterbell10, @stadlmax, is this expected to work or is this a type not handled by this change?